Change parameter data type at embedPunishment() method.
* Change the `duration` parameter in the `embedPunishment()` method to `long`, as Duration objects usually return the seconds in `long`.

+++ src/main/java/ofc/bot/util/embeds/EmbedFactory.java
@@ -171,7 +171,7 @@ public final class EmbedFactory {
                 .build();
     }
 
-    public static MessageEmbed embedPunishment(User user, PunishmentType action, String reason, int duration) {
+    public static MessageEmbed embedPunishment(User user, PunishmentType action, String reason, long duration) {
         OficinaEmbed embed = new OficinaEmbed();
         String header = String.format("%s foi %s", user.getName(), action.getDisplay());
